pyproject.toml
[project]
name = "trmk-blog-generator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "jinja2",
    "markdown",
    "pyyaml",
]
